refactor: UIImage extension 파일 분리
