+++ SOPT-Project/SOPT-Project/week3-project/Model/RecommendModel.swift
@@ -15,21 +15,6 @@ struct Recommend: Hashable {
     let icon: UIImage
 }
 
-extension UIImage {
-    static func coloredImage(
-        color: UIColor,
-        size: CGSize = CGSize(width: 350, height: 250)
-    ) -> UIImage {
-            let rect = CGRect(origin: .zero, size: size)
-            UIGraphicsBeginImageContextWithOptions(size, false, 0)
-            color.setFill()
-            UIRectFill(rect)
-            let image = UIGraphicsGetImageFromCurrentImageContext()
-            UIGraphicsEndImageContext()
-            return image ?? UIImage()
-        }
-}
-
 extension Recommend {
     static let mockData: [Recommend] = [
         Recommend(
